requirements.txt
streamlit == 0.84.0
numpy == 1.19.5
tensorflow == 2.5.0
Pillow == 8.3.0